requirements.txt
Flask
flask-cors
pyodbc
bcrypt
python-dotenv
Werkzeug
datetime
